refactor: flat folders structure
